requirements.txt
tensorflow
numpy
pillow
matplotlib
scikit-learn
